Refactor merge config.rs into commands/config.rs
