+++ sea-orm-cli/src/lib.rs
@@ -1,9 +1,7 @@
 #[cfg(feature = "cli")]
 pub mod cli;
 pub mod commands;
-pub mod config;
 
 #[cfg(feature = "cli")]
 pub use cli::*;
 pub use commands::*;
-pub use config::*;
